fix progress page metadata
